adding makefile and splitting source code into files
removing target from source
